pyproject.toml
[project]
name = "ex1"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "diffusers[torch]",
    "transformers",
    "accelerate",
    "peft",
    "sentencepiece>=0.2.1",
    "protobuf>=7.34.1",
    "bing-image-downloader>=1.1.2",
    "ddgs>=9.6.0",
    "requests>=2.32.5",
    "pillow>=11.3.0",
    # Training utilities
    "bitsandbytes>=0.43.0",
    "tensorboard>=2.16.0",
    "datasets>=2.19.0",
    # Evaluation
    "torch-fidelity>=0.3.0",
    "torchmetrics[multimodal]>=1.3.0",
    "python-dotenv>=1.2.2",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }
